+++ .vscode/settings.json
@@ -9,6 +9,16 @@
         "Reingold",
         "relayout",
         "unawaited",
+        "unwatch",
         "vsync"
-    ]
+    ],
+    "dart.flutterSdkPath": ".fvm/flutter_sdk",
+    // Remove .fvm files from search
+    "search.exclude": {
+        "**/.fvm": true
+    },
+    // Remove from file watching
+    "files.watcherExclude": {
+        "**/.fvm": true
+    }
 }
