+++ examples/flutter_counter/.gitignore
@@ -41,3 +41,5 @@ app.*.map.json
 /android/app/debug
 /android/app/profile
 /android/app/release
+
+.fvm/flutter_sdk
