+++ packages/signals/lib/signals_flutter.dart
@@ -5,4 +5,4 @@ export 'src/extensions/signal.dart';
 export 'src/extensions/value_listenable.dart';
 export 'src/extensions/value_notifier.dart';
 export 'src/watch/widget.dart';
-export 'src/watch/extension.dart' hide getSubscriberCount;
+export 'src/watch/extension.dart' hide getSignalListenSubscriberCount, getSignalWatchSubscriberCount;
