+++ packages/signals/pubspec.yaml
@@ -3,7 +3,7 @@ description: "Reactivity made simple. Do more by doing less. Supports Flutter an
 repository: https://github.com/rodydavis/signals.dart
 homepage: https://rodydavis.github.io/signals.dart
 documentation: https://rodydavis.github.io/signals.dart
-version: 1.5.10
+version: 1.5.11
 
 environment:
   sdk: ">=3.0.0 <4.0.0"
