+++ packages/signals_devtools_extension/.fvm/fvm_config.json
@@ -1,0 +1,4 @@
+{
+  "flutterSdkVersion": "beta",
+  "flavors": {}
+}
