+++ pubspec.yaml
@@ -1,0 +1,7 @@
+name: signals_workspace
+
+environment:
+  sdk: '>=3.0.0 <4.0.0'
+
+dev_dependencies:
+  melos: ^3.4.0
